Let users move Plugins based volume panels to the left
Most of the code to make this work was already here and implemented well by it's author.
The code had no way to know when someone wants it to be on left so I suppose this was a to-
do features which the devs didnt get to work on. I defined a simple config overlay which
you can add in your trees as <bool name="config_audioPanelOnLeftSide">true</bool> and
make it all work.